feat(core): core table descriptors
